app/requirements.txt
fastapi
uvicorn
debugpy
ruff
httpx
pydantic
pydantic_settings
pytest
pytest-xdist
asyncpg
databases
pylint
mypy
doublex
schema
strawberry-graphql[fastapi]
pytest-asyncio
sqlalchemy
aio-pika
faker
